requirements.txt
# =============================================================================
# MedVQA-PEFT 项目依赖文件 (requirements.txt)
# 版本：2026.04
# 项目：Qwen3.5-9B + Gemma-4-E4B-IT + LoRA/DoRA/PiSSA + VQA-RAD/SLAKE
# =============================================================================

# ==================== 【1. Unsloth 核心加速】（必须优先安装） ====================
# Unsloth 是本项目核心加速库，支持 Qwen3.5 和 Gemma-4 vision fine-tuning,安装后会自动包含最新 torch 兼容版本
unsloth==2026.4.6
unsloth-zoo==2026.4.8
# ==================== 【2. PyTorch 基础环境】 ====================
torch>=2.4.0,<2.11.0  # 与 unsloth/unsloth-zoo 要求保持一致
torchvision
torchaudio

# ==================== 【3. Hugging Face 核心生态】 ====================
transformers>=4.51.3,<=5.5.0
peft
trl>=0.18.2,<=0.24.0,!=0.19.0
accelerate
datasets>=3.4.1,<4.4.0
huggingface_hub
bitsandbytes        # 4-bit 量化必需
einops

# ==================== 【4. 实验追踪与可视化】 ====================
wandb
tqdm
rich
pandas
numpy

# ==================== 【5. 评估指标专用】 ====================
openai               # GPT-5-as-a-Judge 语义评分
scikit-learn           # 用于计算 Precision/Recall/F1（可选）
Pillow              # 图像处理

# ==================== 【6. 中文关键词提取（开放式问题 Keyword Recall 用）】 ====================
jieba                 # 中文关键词提取（VQA-RAD/SLAKE 中文场景推荐）
scispacy               # 可选：医学专业实体识别（英文场景更准）

# ==================== 【7. 可选加速与调试工具】 ====================
 flash-attn          # 速度提升（可选，需 CUDA 12.1+）
 xformers            # 内存优化（可选）